+++ test/k8s_sandbox/test_helm.py
@@ -15,7 +15,7 @@ from k8s_sandbox._helm import (
     get_all_release_names,
     uninstall,
 )
-from k8s_sandbox._kubernetes_api import get_current_context_namespace
+from k8s_sandbox._kubernetes_api import get_default_namespace
 
 
 @pytest.fixture
@@ -24,6 +24,7 @@ def uninstallable_release() -> Release:
         __file__,
         chart_path=Path("/non_existent_chart"),
         values_source=ValuesSource.none(),
+        context_name=None,
     )
 
 
@@ -47,7 +48,7 @@ async def test_helm_install_error(
 
 
 async def test_cancelling_install_uninstalls():
-    release = Release(__file__, None, ValuesSource.none())
+    release = Release(__file__, None, ValuesSource.none(), None)
     with patch("k8s_sandbox._helm.uninstall", wraps=uninstall) as spy:
         task = asyncio.create_task(release.install())
         await asyncio.sleep(0.5)
@@ -58,14 +59,14 @@ async def test_cancelling_install_uninstalls():
 
     assert spy.call_count == 1
     assert release.release_name not in await get_all_release_names(
-        get_current_context_namespace()
+        get_default_namespace(context_name=None), None
     )
 
 
 async def test_helm_uninstall_does_not_error_for_release_not_found(
     log_err: LogCaptureFixture,
 ) -> None:
-    release = Release(__file__, None, ValuesSource.none())
+    release = Release(__file__, None, ValuesSource.none(), None)
 
     # Note: we haven't called install() on release.
     await release.uninstall(quiet=False)
@@ -77,7 +78,7 @@ async def test_helm_uninstall_errors_for_other_errors(
     log_err: LogCaptureFixture,
 ) -> None:
     with pytest.raises(RuntimeError) as excinfo:
-        await uninstall("my invalid release name!", "fake-namespace", quiet=False)
+        await uninstall("my invalid release name!", "fake-namespace", None, quiet=False)
 
     assert "Release name is invalid" in log_err.text
     assert "Release name is invalid" in str(excinfo.value)
@@ -116,7 +117,7 @@ async def test_invalid_helm_timeout(
 
 async def test_helm_install_timeout(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setenv(INSPECT_HELM_TIMEOUT, "1")
-    release = Release(__file__, None, ValuesSource.none())
+    release = Release(__file__, None, ValuesSource.none(), None)
 
     with pytest.raises(RuntimeError) as excinfo:
         await release.install()